move creation of new schedule  to the init component
